added projective data type and evaluable class.
The projective data type represents the projectivisation of a field. I've defined field ops, but they are effectively only keeping track of  poles and zeroes to the first order. In particular infty*infty / zero = 1. It probably works well enough. There are two directions to go. 1) remove the field ops and keep it as a purely geometric space. 2) implement it as the space of laurent polynomials (finite orders at zero and infinity)

The Evaluable class is a goldmine. Unifies both evalP and evalPS for polys and series respectively. Wrote an instance for ratio of evaluables, eg rational functions. the 32156 operator works a treat as well, better for both evaluation and partial application (eg making a poly into a function). So p is a poly that can be shown, and p32156 is a function ready to go. Functional dependency used to aid type inference.