Added primes via a library import. Added a bunch of special functions. Added rewrite rules to Integral. Renamed RatioTest from QuotientTest.
The primes works well. The special functions mostly work well, though the zeta function could use some work; it is especially slow near the pole and seems not to converge on the line Re z = 1.

The ContinuedFractions module is very broken, but interesting.

The rewrite rules were finally added to Integral, which should speed up some ops. However, I couldn't get the primitive conversions based on GHC.Integer.Type to work, as there was an error on building.